comment: 불필요한 주석 제거
