+++ src/controllers/LottoDrawing.js
@@ -13,12 +13,10 @@ class LottoDrawing {
     this.#getRandomNumbers = () => Random.pickUniqueNumbersInRange(1, 45, 6);
   }
 
-  //  로또 구입 수량 체크
   checkCount() {
     return Number(this.#line) / 1000;
   }
 
-  // 수량만큼 발행  //TODO: 오름차순정리 리팩토링
   makeLottos() {
     for (let i = 0; i < this.checkCount(); i += 1) {
       this.#lottos.push(modifiers.composeAscending(this.#getRandomNumbers()));
@@ -28,5 +26,3 @@ class LottoDrawing {
 }
 
 export default LottoDrawing;
-// const drawing = new LottoDrawing('3000');
-// console.log(drawing.makeLottos());
