+++ src/controllers/WinnerMaker.js
@@ -3,7 +3,7 @@ import Winner from '../models/Winner.js';
 import modifiers from '../utils/modifiers.js';
 
 class WinnerMaker {
-  #lottos; // 발행한 로또 받기
+  #lottos;
 
   #winner;
 
@@ -15,9 +15,6 @@ class WinnerMaker {
     this.#winningLotto = new Lotto(numbers, bonus);
   }
 
-  // #winningLotto = new Lotto('1,2,3,4,5,12', '7'); // 로또값 입력
-
-  // 일치 개수 확인
   #checkNumbers(lotto) {
     const duplicateNumber = modifiers.getDuplicates(
       lotto,
@@ -26,12 +23,10 @@ class WinnerMaker {
     return duplicateNumber.length;
   }
 
-  // 보너스 확인
   #checkBonus(lotto) {
     return modifiers.isDuplicate(lotto, this.#winningLotto.getBonus());
   }
 
-  // 등수 저장
   #getWinner() {
     this.#lottos.forEach((lotto) =>
       this.#winner.switchNumber(
@@ -50,13 +45,3 @@ class WinnerMaker {
 }
 
 export default WinnerMaker;
-
-// const winnerMaker = new WinnerMaker(
-//   [
-//     [1, 2, 3, 4, 39, 44],
-//     [1, 2, 3, 0, 4, 5],
-//   ],
-//   '1,2,3,0,39,6',
-//   '7',
-// );
-// console.log(winnerMaker.checkNumbers([3, 1, 17, 30, 43, 45]));
