+++ src/models/constants.js
@@ -3,4 +3,14 @@ const USER_INPUT = {
   winningNumbers: '당첨 번호를 입력해 주세요.',
   bonusNumber: '보너스 번호를 입력해 주세요.',
 };
-export default USER_INPUT;
+
+const PRICE = {
+  first: 2000000000,
+  second: 30000000,
+  third: 1500000,
+  forth: 50000,
+  fifth: 5000,
+};
+
+const CONSTANTS = { USER_INPUT, PRICE };
+export default CONSTANTS;
